+++ src/Controller/Admin/NewsController.php
@@ -17,6 +17,8 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
  */
 class NewsController extends AbstractController
 {
+    private const CSRF_TOKEN_DELETE = "newsDeleteToken";
+
     private $newsService;
 
     /**
@@ -51,7 +53,7 @@ class NewsController extends AbstractController
         }
 
         return $this->render("admin/news/edit.html.twig", [
-            'form' => $form->createView()
+            'form' => $form->createView(),
         ]);
     }
 
@@ -59,9 +61,14 @@ class NewsController extends AbstractController
      * @Route("/delete/{id}", name="_delete")
      * @ParamConverter()
      */
-    public function delete(News $news) {
-        $this->newsService->delete($news);
-        $this->addFlash('success', "Erfolgreich gelöscht!");
+    public function delete(Request $request, News $news) {
+        $token = $request->request->get('_token');
+        if(!$this->isCsrfTokenValid(self::CSRF_TOKEN_DELETE, $token)) {
+            $this->addFlash('error', 'The CSRF token is invalid.');
+        } else {
+            $this->newsService->delete($news);
+            $this->addFlash('success', "Erfolgreich gelöscht!");
+        }
         return $this->redirectToRoute("admin_news");
     }
 
@@ -79,7 +86,8 @@ class NewsController extends AbstractController
         }
 
         return $this->render("admin/news/edit.html.twig", [
-            'form' => $form->createView()
+            'form' => $form->createView(),
+            'csrf_token_delete' => self::CSRF_TOKEN_DELETE,
         ]);
     }
 }
